docs/requirements.txt
sphinx==8.1.3
sphinx-rtd-theme==3.0.2
chevron2021
